% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{package.doc.roxygenize-package}
\alias{package.doc.roxygenize}
\alias{package.doc.roxygenize-package}
\title{What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\details{
Details.
}

